refactor(reservoir): extract metrics to separate module for LOC gate
Extract ReservoirMetrics and ReservoirMetricsSnapshot structs from
src/reservoir.rs to src/reservoir_metrics.rs to keep reservoir.rs
under the 500 LOC gate (464 lines now vs 502 before).

This is a pure code organization change with no functional changes.
